+++ Searching Algo/binary_search_using_vector.cpp
